Startup code for shiny app

Prepare inputs for safetyGraphics app - run before app is initialized.

app_startup(
  domainData = NULL,
  meta = NULL,
  charts = NULL,
  mapping = NULL,
  autoMapping = NULL,
  filterDomain = NULL,
  chartSettingsPaths = NULL
)

Arguments

domainData
named list of data.frames to be loaded in to the app. Sample AdAM data from the safetyData package used by default

meta
data frame containing the metadata for use in the app. See the preloaded file (?safetyGraphics::meta) for more data specifications and details. Defaults to safetyGraphics::meta.

charts
list of charts in the format produced by safetyGraphics::makeChartConfig()

mapping
list specifying the initial mapping values for each data mapping for each domain (e.g. list(aes= list(id_col='USUBJID', seq_col='AESEQ')).

autoMapping
boolean indicating whether the app should attempt to automatically detect data standards and generate mappings for the data provided. Values specified in the mapping parameter overwrite automatically generated mappings when both are found. Defaults to true.

filterDomain
domain used for the data/filter tab. Demographics ("dm") is used by default. Using a domain that is not one record per participant is not recommended.

chartSettingsPaths
path(s) where customization functions are saved relative to your working directory. All charts can have itialization (e.g. myChart_Init.R) and static charts can have charting functions (e.g. myGraphic_Chart.R). All R files in this folder are sourced and files with the correct naming convention are linked to the chart. See the Custom Charts vignette for more details.

Value

List of elements for used to initialize the shiny app with the following parameters

"meta" List of configuration metadata

"charts" List of charts

"domainData" List of domain level data sets

"mapping" Initial Data Mapping

"standards" List of domain level data standards