Generate a settings object based on a data standard

This function returns a settings object for the eDish chart based on the specified data standard.

generateSettings(standard = "None", charts = NULL,
  useDefaults = TRUE, partial = FALSE, partial_keys = NULL,
  custom_settings = NULL)

Arguments

standard
The data standard for which to create settings. Valid options are "sdtm", "adam" or "none". Default: "None".

charts
The chart or charts for which settings should be generated. Default: NULL (uses all available charts).

useDefaults
Specifies whether default values from settingsMetadata should be included in the settings object. Default: TRUE.

partial
Boolean for whether or not the standard is a partial standard. Default: FALSE.

partial_keys
Optional character vector of the matched settings if partial is TRUE. Settings should be identified using the text_key format described in ?settingsMetadata. Setting is ignored when partial is FALSE. Default: NULL.

custom_settings
A tibble describing custom settings to be added to the settings object. Custom values overwrite default values when provided. Tibble should have text_key and customValue columns. Default: NULL.

Value

A list containing the appropriate settings for the selected chart

Details

The function is designed to work with the SDTM and ADaM CDISC(<https://www.cdisc.org/>) standards for clinical trial data. Currently, eDish is the only chart supported.

Examples

generateSettings(standard="SDTM")

#> $id_col
#> [1] "USUBJID"
#> 
#> $value_col
#> [1] "STRESN"
#> 
#> $measure_col
#> [1] "TEST"
#> 
#> $measure_values
#> $measure_values$ALT
#> [1] "Aminotransferase, alanine (ALT)"
#> 
#> $measure_values$AST
#> [1] "Aminotransferase, aspartate (AST)"
#> 
#> $measure_values$TB
#> [1] "Total Bilirubin"
#> 
#> $measure_values$ALP
#> [1] "Alkaline phosphatase (ALP)"
#> 
#> 
#> $normal_col_low
#> [1] "STNRLO"
#> 
#> $normal_col_high
#> [1] "STNRHI"
#> 
#> $studyday_col
#> [1] "DY"
#> 
#> $visit_col
#> [1] "VISIT"
#> 
#> $visitn_col
#> [1] "VISITNUM"
#> 
#> $filters
#> NULL
#> 
#> $group_cols
#> NULL
#> 
#> $baseline
#> $baseline$value_col
#> NULL
#> 
#> $baseline$values
#> NULL
#> 
#> 
#> $analysisFlag
#> $analysisFlag$value_col
#> NULL
#> 
#> $analysisFlag$values
#> NULL
#> 
#> 
#> $x_options
#> [1] "ALT" "AST" "ALP"
#> 
#> $y_options
#> [1] "TB"  "ALP"
#> 
#> $visit_window
#> [1] 30
#> 
#> $r_ratio_filter
#> [1] TRUE
#> 
#> $r_ratio_cut
#> [1] 0
#> 
#> $showTitle
#> [1] TRUE
#> 
#> $warningText
#> [1] "Caution: This interactive graphic is not validated. Any clinical recommendations based on this tool should be confirmed using your organizations standard operating procedures."
#> 
#> $unit_col
#> NULL
#> 
#> $start_value
#> NULL
#> 
#> $details
#> NULL
#> 
#> $missingValues
#> [1] ""    "NA"  "N/A"
#> 
#> $visits_without_data
#> [1] FALSE
#> 
#> $unscheduled_visits
#> [1] FALSE
#> 
#> $unscheduled_visit_pattern
#> [1] "/unscheduled|early termination/i"
#> 
#> $calculate_palt
#> NULL
#> 

generateSettings(standard="SdTm") #also ok

#> $id_col
#> [1] "USUBJID"
#> 
#> $value_col
#> [1] "STRESN"
#> 
#> $measure_col
#> [1] "TEST"
#> 
#> $measure_values
#> $measure_values$ALT
#> [1] "Aminotransferase, alanine (ALT)"
#> 
#> $measure_values$AST
#> [1] "Aminotransferase, aspartate (AST)"
#> 
#> $measure_values$TB
#> [1] "Total Bilirubin"
#> 
#> $measure_values$ALP
#> [1] "Alkaline phosphatase (ALP)"
#> 
#> 
#> $normal_col_low
#> [1] "STNRLO"
#> 
#> $normal_col_high
#> [1] "STNRHI"
#> 
#> $studyday_col
#> [1] "DY"
#> 
#> $visit_col
#> [1] "VISIT"
#> 
#> $visitn_col
#> [1] "VISITNUM"
#> 
#> $filters
#> NULL
#> 
#> $group_cols
#> NULL
#> 
#> $baseline
#> $baseline$value_col
#> NULL
#> 
#> $baseline$values
#> NULL
#> 
#> 
#> $analysisFlag
#> $analysisFlag$value_col
#> NULL
#> 
#> $analysisFlag$values
#> NULL
#> 
#> 
#> $x_options
#> [1] "ALT" "AST" "ALP"
#> 
#> $y_options
#> [1] "TB"  "ALP"
#> 
#> $visit_window
#> [1] 30
#> 
#> $r_ratio_filter
#> [1] TRUE
#> 
#> $r_ratio_cut
#> [1] 0
#> 
#> $showTitle
#> [1] TRUE
#> 
#> $warningText
#> [1] "Caution: This interactive graphic is not validated. Any clinical recommendations based on this tool should be confirmed using your organizations standard operating procedures."
#> 
#> $unit_col
#> NULL
#> 
#> $start_value
#> NULL
#> 
#> $details
#> NULL
#> 
#> $missingValues
#> [1] ""    "NA"  "N/A"
#> 
#> $visits_without_data
#> [1] FALSE
#> 
#> $unscheduled_visits
#> [1] FALSE
#> 
#> $unscheduled_visit_pattern
#> [1] "/unscheduled|early termination/i"
#> 
#> $calculate_palt
#> NULL
#> 

generateSettings(standard="ADaM")

#> $id_col
#> [1] "USUBJID"
#> 
#> $value_col
#> [1] "AVAL"
#> 
#> $measure_col
#> [1] "PARAM"
#> 
#> $measure_values
#> $measure_values$ALT
#> [1] "Alanine Aminotransferase (U/L)"
#> 
#> $measure_values$AST
#> [1] "Aspartate Aminotransferase (U/L)"
#> 
#> $measure_values$TB
#> [1] "Bilirubin (umol/L)"
#> 
#> $measure_values$ALP
#> [1] "Alkaline Phosphatase (U/L)"
#> 
#> 
#> $normal_col_low
#> [1] "A1LO"
#> 
#> $normal_col_high
#> [1] "A1HI"
#> 
#> $studyday_col
#> [1] "ADY"
#> 
#> $visit_col
#> [1] "VISIT"
#> 
#> $visitn_col
#> [1] "VISITNUM"
#> 
#> $filters
#> NULL
#> 
#> $group_cols
#> NULL
#> 
#> $baseline
#> $baseline$value_col
#> NULL
#> 
#> $baseline$values
#> NULL
#> 
#> 
#> $analysisFlag
#> $analysisFlag$value_col
#> NULL
#> 
#> $analysisFlag$values
#> NULL
#> 
#> 
#> $x_options
#> [1] "ALT" "AST" "ALP"
#> 
#> $y_options
#> [1] "TB"  "ALP"
#> 
#> $visit_window
#> [1] 30
#> 
#> $r_ratio_filter
#> [1] TRUE
#> 
#> $r_ratio_cut
#> [1] 0
#> 
#> $showTitle
#> [1] TRUE
#> 
#> $warningText
#> [1] "Caution: This interactive graphic is not validated. Any clinical recommendations based on this tool should be confirmed using your organizations standard operating procedures."
#> 
#> $unit_col
#> NULL
#> 
#> $start_value
#> NULL
#> 
#> $details
#> NULL
#> 
#> $missingValues
#> [1] ""    "NA"  "N/A"
#> 
#> $visits_without_data
#> [1] FALSE
#> 
#> $unscheduled_visits
#> [1] FALSE
#> 
#> $unscheduled_visit_pattern
#> [1] "/unscheduled|early termination/i"
#> 
#> $calculate_palt
#> NULL
#> 

pkeys<- c("id_col","measure_col","value_col")
generateSettings(standard="adam", partial=TRUE, partial_keys=pkeys)

#> $id_col
#> [1] "USUBJID"
#> 
#> $value_col
#> [1] "AVAL"
#> 
#> $measure_col
#> [1] "PARAM"
#> 
#> $measure_values
#> $measure_values$ALT
#> NULL
#> 
#> $measure_values$AST
#> NULL
#> 
#> $measure_values$TB
#> NULL
#> 
#> $measure_values$ALP
#> NULL
#> 
#> 
#> $normal_col_low
#> NULL
#> 
#> $normal_col_high
#> NULL
#> 
#> $studyday_col
#> NULL
#> 
#> $visit_col
#> NULL
#> 
#> $visitn_col
#> NULL
#> 
#> $filters
#> NULL
#> 
#> $group_cols
#> NULL
#> 
#> $baseline
#> $baseline$value_col
#> NULL
#> 
#> $baseline$values
#> NULL
#> 
#> 
#> $analysisFlag
#> $analysisFlag$value_col
#> NULL
#> 
#> $analysisFlag$values
#> NULL
#> 
#> 
#> $x_options
#> [1] "ALT" "AST" "ALP"
#> 
#> $y_options
#> [1] "TB"  "ALP"
#> 
#> $visit_window
#> [1] 30
#> 
#> $r_ratio_filter
#> [1] TRUE
#> 
#> $r_ratio_cut
#> [1] 0
#> 
#> $showTitle
#> [1] TRUE
#> 
#> $warningText
#> [1] "Caution: This interactive graphic is not validated. Any clinical recommendations based on this tool should be confirmed using your organizations standard operating procedures."
#> 
#> $unit_col
#> NULL
#> 
#> $start_value
#> NULL
#> 
#> $details
#> NULL
#> 
#> $missingValues
#> [1] ""    "NA"  "N/A"
#> 
#> $visits_without_data
#> [1] FALSE
#> 
#> $unscheduled_visits
#> [1] FALSE
#> 
#> $unscheduled_visit_pattern
#> [1] "/unscheduled|early termination/i"
#> 
#> $calculate_palt
#> NULL
#> 

generateSettings(standard="a different standard")

#> $id_col
#> NULL
#> 
#> $value_col
#> NULL
#> 
#> $measure_col
#> NULL
#> 
#> $measure_values
#> $measure_values$ALT
#> NULL
#> 
#> $measure_values$AST
#> NULL
#> 
#> $measure_values$TB
#> NULL
#> 
#> $measure_values$ALP
#> NULL
#> 
#> 
#> $normal_col_low
#> NULL
#> 
#> $normal_col_high
#> NULL
#> 
#> $studyday_col
#> NULL
#> 
#> $visit_col
#> NULL
#> 
#> $visitn_col
#> NULL
#> 
#> $filters
#> NULL
#> 
#> $group_cols
#> NULL
#> 
#> $baseline
#> $baseline$value_col
#> NULL
#> 
#> $baseline$values
#> NULL
#> 
#> 
#> $analysisFlag
#> $analysisFlag$value_col
#> NULL
#> 
#> $analysisFlag$values
#> NULL
#> 
#> 
#> $x_options
#> [1] "ALT" "AST" "ALP"
#> 
#> $y_options
#> [1] "TB"  "ALP"
#> 
#> $visit_window
#> [1] 30
#> 
#> $r_ratio_filter
#> [1] TRUE
#> 
#> $r_ratio_cut
#> [1] 0
#> 
#> $showTitle
#> [1] TRUE
#> 
#> $warningText
#> [1] "Caution: This interactive graphic is not validated. Any clinical recommendations based on this tool should be confirmed using your organizations standard operating procedures."
#> 
#> $unit_col
#> NULL
#> 
#> $start_value
#> NULL
#> 
#> $details
#> NULL
#> 
#> $missingValues
#> [1] ""    "NA"  "N/A"
#> 
#> $visits_without_data
#> [1] FALSE
#> 
#> $unscheduled_visits
#> [1] FALSE
#> 
#> $unscheduled_visit_pattern
#> [1] "/unscheduled|early termination/i"
#> 
#> $calculate_palt
#> NULL
#> 

#returns shell settings list with no data mapping