Generate a default settings shell based on settings metadata

This function returns a default settings object based on the chart(s) specified.

generateShell(charts = NULL)

Arguments

charts
The chart or chart(s) to include in the shell settings object

Value

A list containing a setting shell (all values = NA) for the selected chart(s)

Details

The function is designed to work with valid safetyGraphics charts.

Examples

safetyGraphics:::generateShell(charts = "hepexplorer")

#> $id_col
#> NULL
#> 
#> $value_col
#> NULL
#> 
#> $measure_col
#> NULL
#> 
#> $measure_values
#> $measure_values$ALT
#> NULL
#> 
#> $measure_values$AST
#> NULL
#> 
#> $measure_values$TB
#> NULL
#> 
#> $measure_values$ALP
#> NULL
#> 
#> 
#> $normal_col_low
#> NULL
#> 
#> $normal_col_high
#> NULL
#> 
#> $studyday_col
#> NULL
#> 
#> $visit_col
#> NULL
#> 
#> $visitn_col
#> NULL
#> 
#> $filters
#> NULL
#> 
#> $group_cols
#> NULL
#> 
#> $baseline
#> $baseline$value_col
#> NULL
#> 
#> $baseline$values
#> NULL
#> 
#> 
#> $analysisFlag
#> $analysisFlag$value_col
#> NULL
#> 
#> $analysisFlag$values
#> NULL
#> 
#> 
#> $x_options
#> NULL
#> 
#> $y_options
#> NULL
#> 
#> $visit_window
#> NULL
#> 
#> $r_ratio_filter
#> NULL
#> 
#> $r_ratio_cut
#> NULL
#> 
#> $showTitle
#> NULL
#> 
#> $warningText
#> NULL
#> 
#> $details
#> NULL
#>