Get setting keys matching a pattern

Recursive function to find all keys matching a given text pattern in a settings list

getSettingKeys(patterns, settings, parents = NULL, matchLists = FALSE)

Arguments

patterns
List of text patterns to match with named elements in settings.

settings
List of settings used to generate a chart like eDISH().

parents
List containing the position of the parent list using recursive matches.

matchLists
Boolean indicating whether keys containing lists should be returned as matches.

Value

List of lists specifying the position of matching named elements in the format list("filters",2,"value_col"), which would correspond to settings[["filters"]][[2]][["value_col"]].

Details

This function loops through all named elements (or "keys") in a settings list and returns items that match patterns. If matchLists==FALSE (the default), the function iteratively looks through the named elements in nested lists using the built-in parents parameter. The function returns a array of keys for all matches using a list of lists. Each key is defines the position of a matching key using an unnamed list. For example, list("filters",2,"value_col") would correspond to settings[["filters"]][[2]][["value_col"]].

Examples

testSettings<-generateSettings(standard="AdAM")

# returns list of all matching values
safetyGraphics:::getSettingKeys(patterns=c("_col"),
                                settings=testSettings)

#> [[1]]
#> [[1]][[1]]
#> [1] "id_col"
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] "value_col"
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#> [1] "measure_col"
#> 
#> 
#> [[4]]
#> [[4]][[1]]
#> [1] "normal_col_low"
#> 
#> 
#> [[5]]
#> [[5]][[1]]
#> [1] "normal_col_high"
#> 
#> 
#> [[6]]
#> [[6]][[1]]
#> [1] "studyday_col"
#> 
#> 
#> [[7]]
#> [[7]][[1]]
#> [1] "visit_col"
#> 
#> 
#> [[8]]
#> [[8]][[1]]
#> [1] "visitn_col"
#> 
#> 
#> [[9]]
#> [[9]][[1]]
#> [1] "group_cols"
#> 
#> 
#> [[10]]
#> [[10]][[1]]
#> [1] "baseline"
#> 
#> [[10]][[2]]
#> [1] "value_col"
#> 
#> 
#> [[11]]
#> [[11]][[1]]
#> [1] "analysisFlag"
#> 
#> [[11]][[2]]
#> [1] "value_col"
#> 
#> 
#> [[12]]
#> [[12]][[1]]
#> [1] "unit_col"
#> 
#> 

#finds the matching nested setting 
safetyGraphics:::getSettingKeys(patterns=c("ALP"),
                                settings=testSettings)

#> [[1]]
#> [[1]][[1]]
#> [1] "measure_values"
#> 
#> [[1]][[2]]
#> [1] "ALP"
#> 
#> 

#returns an empty list, since the only matching item is a list
safetyGraphics:::getSettingKeys(patterns=c("measure_values"),
                                settings=testSettings)

#> list()

#finds the matching key associated with a list
safetyGraphics:::getSettingKeys(patterns=c("measure_values"),
                                settings=testSettings,
                                matchLists=TRUE)

#> [[1]]
#> [[1]][[1]]
#> [1] "measure_values"
#> 
#>