Get metadata about chart settings

Retrieve specified metadata about chart settings from the data/settingsMetadata.Rda file.

getSettingsMetadata(charts = NULL, text_keys = NULL, cols = NULL,
  filter_expr = NULL, add_standards = TRUE,
  metadata = safetyGraphics::settingsMetadata)

Arguments

charts
optional vector of chart names used to filter the metadata. Exact matches only (case-insensitive). All rows returned by default.

text_keys
optional vector of keys used to filter the metadata. Partial matches for any of the strings are returned (case-insensitive). All rows returned by default.

cols
optional vector of columns to return from the metadata. All columns returned by default.

filter_expr
optional filter expression used to subset the data.

add_standards
should data standard info stored in standardsMetadata be included

metadata
metadata data frame to be queried

Value

dataframe with the requested metadata or single metadata value

Examples

safetyGraphics:::getSettingsMetadata()

#>                     text_key                            label
#> 1                     id_col                        ID column
#> 2                  value_col                     Value column
#> 3                measure_col                   Measure column
#> 4        measure_values--ALT   Alanine Aminotransferase value
#> 5        measure_values--AST Aspartate Aminotransferase value
#> 6         measure_values--TB            Total Bilirubin value
#> 7        measure_values--ALP       Alkaline Phosphatase value
#> 8             normal_col_low     Lower Limit of Normal column
#> 9            normal_col_high     Upper Limit of Normal column
#> 10              studyday_col                 Study Day column
#> 11                 visit_col                     Visit column
#> 12                visitn_col              Visit Number column
#> 13                   filters                  Filters columns
#> 14                group_cols                    Group columns
#> 15       baseline--value_col                  Baseline column
#> 16          baseline--values                  Baseline values
#> 17   analysisFlag--value_col             Analysis Flag column
#> 18      analysisFlag--values             Analysis Flag values
#> 19                 x_options                   X axis options
#> 20                 y_options                   Y axis options
#> 21              visit_window    Default Visit Window in Days 
#> 22            r_ratio_filter             Show R Ratio Filter?
#> 23               r_ratio_cut              Default R Ratio Cut
#> 24                 showTitle               Show Chart Title? 
#> 25               warningText                     Warning text
#> 26                  unit_col                      Unit column
#> 27               start_value              Measure start value
#> 28                   details                  Details columns
#> 29             missingValues                   Missing values
#> 30       visits_without_data              Visits without data
#> 31        unscheduled_visits               Unscheduled visits
#> 32 unscheduled_visit_pattern        Unscheduled visit pattern
#>                                                                                                                                                                                                                      description
#> 1                                                                                                                                                                                       Unique subject identifier variable name.
#> 2                                                                                                                                                                                                      Lab result variable name.
#> 3                                                                                                                                                                                                      Lab measure variable name
#> 4                                                                                                                                                        Value used for Alanine Aminotransferase in the specified measure column
#> 5                                                                                                                                                      Value used for Aspartate Aminotransferase in the specified measure column
#> 6                                                                                                                                                                 Value used for Total Bilirubin in the specified measure column
#> 7                                                                                                                                                            Value used for Alkaline Phosphatase in the specified measure column
#> 8                                                                                                                                                                                            Lower limit of normal variable name
#> 9                                                                                                                                                                                            Upper limit of normal variable name
#> 10                                                                                                                                                                                                       Visit day variable name
#> 11                                                                                                                                                                                                           Visit variable name
#> 12                                                                                                                                                                                                    Visit number variable name
#> 13                                   An optional list of specifications for filters. Each filter is a nested, named list (containing the filter value column: 'value_col' and associated label: 'label') within the larger list.
#> 14                 An optional list of specifications for grouping columns. Each group column is a nested, named list (containing the group variable column: 'value_col' and associated label: 'label') within the larger list. 
#> 15                                                                                                                                          An optional list defining which column represent the baseline visit(s) of the study.
#> 16                                                                                                                            An optional list defining which values (one or more) represent the baseline visit(s) of the study.
#> 17                                                                                                                                            An optional list defining which column should be used in eDish and mDish analyses.
#> 18                                                                                                                                            An optional list defining which values should be used in eDish and mDish analyses.
#> 19           Specifies variable options for the x-axis using the key values from measure_values (e.g. 'ALT'). When multiple options are specified, a control allowing the user to interactively change the x variable is shown. 
#> 20            Specifies variable options for the y-axis using the key values from measure_values (e.g. 'TB'). When multiple options are specified, a control allowing the user to interactively change the y variable is shown. 
#> 21                                                                      Default visit window used to highlight eDish points where x and y measures occurred within the specified number of days. Editable by user after render. 
#> 22                                                                                               Specifies whether the R Ratio filter should be shown. R ratio is defined as: (ALT value/ULN for ALT) / (ALP value/ULN for ALP).
#> 23                                                                               Default cut point for R Ratio filter. Ignored when r_ratio_filter = FALSE. User can update this setting via the UI when r_ratio_filter = TRUE. 
#> 24                                                                                                                                                               Specifies whether the title should be drawn above the controls.
#> 25                                                                                    Informational text to be displayed near the top of the controls (beneath the title, if any). No warning is displayed if warningText = ''. 
#> 26                                                                                                                                                                                                 Unit of measure variable name
#> 27                                                                                                                                Value of variable defined in measure_col to be rendered in the histogram when the widget loads
#> 28 An optional list of specifications for details listing.  Each column to be added to details listing is a nested, named list (containing the variable name: "value_col" and associated label: "label") within the larger list.
#> 29                                                                                                                                                                Values defining a missing value in the selected 'value' column
#> 30                                                                                                                                                               Controls display of visits without data for the current measure
#> 31                                                                                                                                                                                        Controls display of unscheduled visits
#> 32                                                                                                                                                                       A regular expression that identifies unscheduled visits
#>    setting_type setting_required column_mapping column_type field_mapping
#> 1     character             TRUE           TRUE   character         FALSE
#> 2     character             TRUE           TRUE     numeric         FALSE
#> 3     character             TRUE           TRUE   character         FALSE
#> 4     character             TRUE          FALSE        <NA>          TRUE
#> 5     character             TRUE          FALSE        <NA>          TRUE
#> 6     character             TRUE          FALSE        <NA>          TRUE
#> 7     character             TRUE          FALSE        <NA>          TRUE
#> 8     character             TRUE           TRUE     numeric         FALSE
#> 9     character             TRUE           TRUE     numeric         FALSE
#> 10    character             TRUE           TRUE     numeric         FALSE
#> 11    character             TRUE           TRUE   character         FALSE
#> 12    character             TRUE           TRUE     numeric         FALSE
#> 13       vector            FALSE           TRUE        <NA>         FALSE
#> 14       vector            FALSE           TRUE        <NA>         FALSE
#> 15    character            FALSE           TRUE        <NA>         FALSE
#> 16       vector            FALSE          FALSE        <NA>          TRUE
#> 17    character            FALSE           TRUE        <NA>         FALSE
#> 18       vector            FALSE          FALSE        <NA>          TRUE
#> 19       vector            FALSE          FALSE        <NA>         FALSE
#> 20       vector            FALSE          FALSE        <NA>         FALSE
#> 21      numeric            FALSE          FALSE        <NA>         FALSE
#> 22      logical            FALSE          FALSE        <NA>         FALSE
#> 23      numeric            FALSE          FALSE        <NA>         FALSE
#> 24      logical            FALSE          FALSE        <NA>         FALSE
#> 25    character            FALSE          FALSE        <NA>         FALSE
#> 26    character            FALSE           TRUE   character         FALSE
#> 27    character            FALSE          FALSE        <NA>          TRUE
#> 28       vector            FALSE           TRUE        <NA>         FALSE
#> 29       vector            FALSE          FALSE        <NA>         FALSE
#> 30      logical            FALSE          FALSE        <NA>         FALSE
#> 31      logical            FALSE          FALSE        <NA>         FALSE
#> 32    character            FALSE          FALSE        <NA>         FALSE
#>           field_column_key setting_cat
#> 1                                 data
#> 2                                 data
#> 3                                 data
#> 4              measure_col        data
#> 5              measure_col        data
#> 6              measure_col        data
#> 7              measure_col        data
#> 8                                 data
#> 9                                 data
#> 10                                data
#> 11                                data
#> 12                                data
#> 13                                data
#> 14                                data
#> 15                                data
#> 16     baseline--value_col        data
#> 17                                data
#> 18 analysisFlag--value_col        data
#> 19                             measure
#> 20                             measure
#> 21                          appearance
#> 22                          appearance
#> 23                          appearance
#> 24                          appearance
#> 25                          appearance
#> 26                                data
#> 27             measure_col        data
#> 28                                data
#> 29                                data
#> 30                                data
#> 31                                data
#> 32                                data
#>                                                                                                                                                                            default
#> 1                                                                                                                                                                             NULL
#> 2                                                                                                                                                                             NULL
#> 3                                                                                                                                                                             NULL
#> 4                                                                                                                                                                             NULL
#> 5                                                                                                                                                                             NULL
#> 6                                                                                                                                                                             NULL
#> 7                                                                                                                                                                             NULL
#> 8                                                                                                                                                                             NULL
#> 9                                                                                                                                                                             NULL
#> 10                                                                                                                                                                            NULL
#> 11                                                                                                                                                                            NULL
#> 12                                                                                                                                                                            NULL
#> 13                                                                                                                                                                            NULL
#> 14                                                                                                                                                                            NULL
#> 15                                                                                                                                                                            NULL
#> 16                                                                                                                                                                            NULL
#> 17                                                                                                                                                                            NULL
#> 18                                                                                                                                                                            NULL
#> 19                                                                                                                                                                   ALT, AST, ALP
#> 20                                                                                                                                                                         TB, ALP
#> 21                                                                                                                                                                              30
#> 22                                                                                                                                                                            TRUE
#> 23                                                                                                                                                                               0
#> 24                                                                                                                                                                            TRUE
#> 25 Caution: This interactive graphic is not validated. Any clinical recommendations based on this tool should be confirmed using your organizations standard operating procedures.
#> 26                                                                                                                                                                            NULL
#> 27                                                                                                                                                                            NULL
#> 28                                                                                                                                                                            NULL
#> 29                                                                                                                                                                       , NA, N/A
#> 30                                                                                                                                                                           FALSE
#> 31                                                                                                                                                                           FALSE
#> 32                                                                                                                                                /unscheduled|early termination/i
#>    chart_hepexplorer chart_safetyhistogram chart_safetyoutlierexplorer
#> 1               TRUE                  TRUE                        TRUE
#> 2               TRUE                  TRUE                        TRUE
#> 3               TRUE                  TRUE                        TRUE
#> 4               TRUE                 FALSE                       FALSE
#> 5               TRUE                 FALSE                       FALSE
#> 6               TRUE                 FALSE                       FALSE
#> 7               TRUE                 FALSE                       FALSE
#> 8               TRUE                  TRUE                        TRUE
#> 9               TRUE                  TRUE                        TRUE
#> 10              TRUE                 FALSE                        TRUE
#> 11              TRUE                 FALSE                        TRUE
#> 12              TRUE                 FALSE                        TRUE
#> 13              TRUE                  TRUE                        TRUE
#> 14              TRUE                 FALSE                       FALSE
#> 15              TRUE                 FALSE                       FALSE
#> 16              TRUE                 FALSE                       FALSE
#> 17              TRUE                 FALSE                       FALSE
#> 18              TRUE                 FALSE                       FALSE
#> 19              TRUE                 FALSE                       FALSE
#> 20              TRUE                 FALSE                       FALSE
#> 21              TRUE                 FALSE                       FALSE
#> 22              TRUE                 FALSE                       FALSE
#> 23              TRUE                 FALSE                       FALSE
#> 24              TRUE                 FALSE                       FALSE
#> 25              TRUE                 FALSE                       FALSE
#> 26             FALSE                  TRUE                        TRUE
#> 27             FALSE                  TRUE                        TRUE
#> 28              TRUE                  TRUE                        TRUE
#> 29             FALSE                  TRUE                       FALSE
#> 30             FALSE                 FALSE                        TRUE
#> 31             FALSE                 FALSE                        TRUE
#> 32             FALSE                 FALSE                        TRUE
#>    chart_safetyshiftplot chart_safetyresultsovertime
#> 1                   TRUE                        TRUE
#> 2                   TRUE                        TRUE
#> 3                   TRUE                        TRUE
#> 4                  FALSE                       FALSE
#> 5                  FALSE                       FALSE
#> 6                  FALSE                       FALSE
#> 7                  FALSE                       FALSE
#> 8                  FALSE                        TRUE
#> 9                  FALSE                        TRUE
#> 10                 FALSE                       FALSE
#> 11                  TRUE                        TRUE
#> 12                 FALSE                        TRUE
#> 13                  TRUE                        TRUE
#> 14                 FALSE                        TRUE
#> 15                 FALSE                       FALSE
#> 16                 FALSE                       FALSE
#> 17                 FALSE                       FALSE
#> 18                 FALSE                       FALSE
#> 19                 FALSE                       FALSE
#> 20                 FALSE                       FALSE
#> 21                 FALSE                       FALSE
#> 22                 FALSE                       FALSE
#> 23                 FALSE                       FALSE
#> 24                 FALSE                       FALSE
#> 25                 FALSE                       FALSE
#> 26                  TRUE                        TRUE
#> 27                  TRUE                        TRUE
#> 28                 FALSE                       FALSE
#> 29                 FALSE                       FALSE
#> 30                 FALSE                        TRUE
#> 31                 FALSE                        TRUE
#> 32                 FALSE                        TRUE
#>    chart_paneledoutlierexplorer chart_safetydeltadelta
#> 1                          TRUE                   TRUE
#> 2                          TRUE                   TRUE
#> 3                          TRUE                   TRUE
#> 4                         FALSE                  FALSE
#> 5                         FALSE                  FALSE
#> 6                         FALSE                  FALSE
#> 7                         FALSE                  FALSE
#> 8                          TRUE                  FALSE
#> 9                          TRUE                  FALSE
#> 10                         TRUE                  FALSE
#> 11                         TRUE                   TRUE
#> 12                         TRUE                   TRUE
#> 13                         TRUE                   TRUE
#> 14                        FALSE                  FALSE
#> 15                        FALSE                  FALSE
#> 16                        FALSE                  FALSE
#> 17                        FALSE                  FALSE
#> 18                        FALSE                  FALSE
#> 19                        FALSE                  FALSE
#> 20                        FALSE                  FALSE
#> 21                        FALSE                  FALSE
#> 22                        FALSE                  FALSE
#> 23                        FALSE                  FALSE
#> 24                        FALSE                  FALSE
#> 25                        FALSE                  FALSE
#> 26                         TRUE                  FALSE
#> 27                        FALSE                  FALSE
#> 28                        FALSE                   TRUE
#> 29                        FALSE                  FALSE
#> 30                         TRUE                  FALSE
#> 31                         TRUE                  FALSE
#> 32                         TRUE                  FALSE
#>                                 sdtm                             adam
#> 1                            USUBJID                          USUBJID
#> 2                             STRESN                             AVAL
#> 3                               TEST                            PARAM
#> 4    Aminotransferase, alanine (ALT)   Alanine Aminotransferase (U/L)
#> 5  Aminotransferase, aspartate (AST) Aspartate Aminotransferase (U/L)
#> 6                    Total Bilirubin               Bilirubin (umol/L)
#> 7         Alkaline phosphatase (ALP)       Alkaline Phosphatase (U/L)
#> 8                             STNRLO                             A1LO
#> 9                             STNRHI                             A1HI
#> 10                                DY                              ADY
#> 11                             VISIT                            VISIT
#> 12                          VISITNUM                         VISITNUM
#> 13                                                                   
#> 14                                                                   
#> 15                                                                   
#> 16                                                                   
#> 17                                                                   
#> 18                                                                   
#> 19                                                                   
#> 20                                                                   
#> 21                                                                   
#> 22                                                                   
#> 23                                                                   
#> 24                                                                   
#> 25                                                                   
#> 26                            STRESU                                 
#> 27                                                                   
#> 28                                                                   
#> 29                                                                   
#> 30                                                                   
#> 31                                                                   
#> 32                                                                   

# Returns a full copy of settingsMetadata.Rda

safetyGraphics:::getSettingsMetadata(text_keys=c("id_col"))

#>   text_key     label                              description setting_type
#> 1   id_col ID column Unique subject identifier variable name.    character
#>   setting_required column_mapping column_type field_mapping field_column_key
#> 1             TRUE           TRUE   character         FALSE                 
#>   setting_cat default chart_hepexplorer chart_safetyhistogram
#> 1        data    NULL              TRUE                  TRUE
#>   chart_safetyoutlierexplorer chart_safetyshiftplot chart_safetyresultsovertime
#> 1                        TRUE                  TRUE                        TRUE
#>   chart_paneledoutlierexplorer chart_safetydeltadelta    sdtm    adam
#> 1                         TRUE                   TRUE USUBJID USUBJID

# returns a dataframe with a single row with metadata for the id_col setting

safetyGraphics:::getSettingsMetadata(text_keys=c("id_col"), cols=c("label"))

#> [1] "ID column"

# returns the character value for the specified row.