Check whether a specified field value is found in a data set

This checks whether a specific value is found in a specified column in a specified data set

hasField(fieldValue, columnName, data)

Arguments

fieldValue
A value to check for.

columnName
The column to check.

data
the data.frame to search.

Value

logical scalar. TRUE if field_value is found. FALSE otherwise

Examples

safetyGraphics:::hasField(fieldValue="Bilirubin (umol/L)",columnName="PARAM",data=adlbc) #TRUE

#> [1] TRUE

safetyGraphics:::hasField(fieldValue="Not_a_real_value",columnName="",data=adlbc) #FALSE

#> [1] FALSE