Create a metadata object table for a set of charts

Generates metadata object for a list of charts. makeMeta() looks for metadata in 3 locations for each chart object:

Domain-level metadata saved as meta_chart$name in the chart$package namespace

Chart-specific metadata saved as meta_chart$domain in the chart$package namespace

Chart-specific metadata saved directly to the chart object as chart$meta After checking all charts, all metadata files are stacked in to a single dataframe and returned. If duplicate metadata rows (domain + text_key) are found, an error is thrown.

makeMeta(charts)

Arguments

charts
list of safetyGraphics chart objects for which to create metadata

Value

tibble of metadata with the following columns:

domain

Data domain

text_key

Text key indicating the setting name. '--' delimiter indicates a field level data mapping

col_key

Key for the column mapping

field_key

Key for the field mapping (if any)

type

type of mapping - "field" or "column"

label

Label

description

Description

multiple

Mapping supports multiple columns/fields

standard_adam

Default values for the ADaM data standard

standard_sdtm

Default values for the SDTM data standard