Helper function to convert keys from text to nested lists

Convert settings keys from text vectors (using the "--" delimiter) to a list of lists

textKeysToList(textKeys)

Arguments

textKeys

a list (or vector) of character keys using the "--" delimiter to indicate hierarchy

Value

A list of unnamed lists, with position in the nested list indicating hierarchy

Examples

safetyGraphics:::textKeysToList("id_col") 
#> [[1]]
#> [[1]][[1]]
#> [1] "id_col"
#> 
#> 
#list(list("id_col"))

#list(list("id_col"),list("measure_col","label"))
safetyGraphics:::textKeysToList(c("id_col","measure_col--label")) 
#> [[1]]
#> [[1]][[1]]
#> [1] "id_col"
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] "measure_col"
#> 
#> [[2]][[2]]
#> [1] "label"
#> 
#>