Compare a settings object with a specified data set

This function returns a list describing the validation status of a data set for a specified data standard

validateSettings(data, settings, charts = NULL)

Arguments

data
A data frame to check against the settings object

settings
The settings list to compare with the data frame.

charts
The charts being validated

Value

A list describing the validation state for the data/settings combination. The returned list has the following properties:

"valid" - boolean indicating whether the settings/data combo is valid for all charts

"status" - string summarizing the validation results

"charts" - a list of lists specifying whether each chart is valid. Each item in the list has "chart" and "valid" properties

"checkList" - list of lists giving details about checks performed on individual setting specifications. Each embedded item has the following properties:

"key" - list specifying the position of the property being checked. For example, `list("group_cols",1,"value_col")` corresponds to `settings[["group_cols"]][[1]][["value_col"]]`

"text_key" - list from `key` parsed to character with a "--" separator.

"value" - value of the setting

"type" - type of the check performed.

"description" - description of the check performed.

"valid" - boolean indicating whether the check was passed

"message" - string describing failed checks (where `valid=FALSE`). returns an empty string when `valid==TRUE`

Details

This function returns a list describing the validation status of a settings/data combo for a given chart type. This list can be used to populate status fields and control workflow in the Shiny app. It could also be used to manually QC a buggy chart. The tool checks that all setting properties containing "_col" match columns in the data set via checkColumnSettings, and all properties containing "_values" match fields in the data set via checkFieldSettings.

Examples

testSettings <- generateSettings(standard="adam")
validateSettings(data=adlbc, settings=testSettings)

#> $checks
#> # A tibble: 33 x 7
#>    key     text_key     type   description           value         valid message
#>    <list>  <chr>        <chr>  <chr>                 <chr>         <lgl> <chr>  
#>  1 <list … id_col       requi… value for specified … USUBJID       TRUE  ""     
#>  2 <list … value_col    requi… value for specified … AVAL          TRUE  ""     
#>  3 <list … measure_col  requi… value for specified … PARAM         TRUE  ""     
#>  4 <list … measure_val… requi… value for specified … Alanine Amin… TRUE  ""     
#>  5 <list … measure_val… requi… value for specified … Aspartate Am… TRUE  ""     
#>  6 <list … measure_val… requi… value for specified … Bilirubin (u… TRUE  ""     
#>  7 <list … measure_val… requi… value for specified … Alkaline Pho… TRUE  ""     
#>  8 <list … normal_col_… requi… value for specified … A1LO          TRUE  ""     
#>  9 <list … normal_col_… requi… value for specified … A1HI          TRUE  ""     
#> 10 <list … studyday_col requi… value for specified … ADY           TRUE  ""     
#> # … with 23 more rows
#> 
#> $valid
#> [1] TRUE
#> 
#> $charts
#> $charts$hepexplorer
#> [1] TRUE
#> 
#> $charts$safetyhistogram
#> [1] TRUE
#> 
#> $charts$safetyoutlierexplorer
#> [1] TRUE
#> 
#> $charts$safetyshiftplot
#> [1] TRUE
#> 
#> $charts$safetyresultsovertime
#> [1] TRUE
#> 
#> $charts$paneledoutlierexplorer
#> [1] TRUE
#> 
#> $charts$safetydeltadelta
#> [1] TRUE
#> 
#> 
#> $status
#> [1] "0 of 33 checks failed."
#> 

# .$valid is TRUE
testSettings$id_col <- "NotAColumn"
validateSettings(data=adlbc, settings=testSettings)

#> $checks
#> # A tibble: 33 x 7
#>    key     text_key     type   description           value         valid message
#>    <list>  <chr>        <chr>  <chr>                 <chr>         <lgl> <chr>  
#>  1 <list … id_col       requi… value for specified … NotAColumn    TRUE  ""     
#>  2 <list … value_col    requi… value for specified … AVAL          TRUE  ""     
#>  3 <list … measure_col  requi… value for specified … PARAM         TRUE  ""     
#>  4 <list … measure_val… requi… value for specified … Alanine Amin… TRUE  ""     
#>  5 <list … measure_val… requi… value for specified … Aspartate Am… TRUE  ""     
#>  6 <list … measure_val… requi… value for specified … Bilirubin (u… TRUE  ""     
#>  7 <list … measure_val… requi… value for specified … Alkaline Pho… TRUE  ""     
#>  8 <list … normal_col_… requi… value for specified … A1LO          TRUE  ""     
#>  9 <list … normal_col_… requi… value for specified … A1HI          TRUE  ""     
#> 10 <list … studyday_col requi… value for specified … ADY           TRUE  ""     
#> # … with 23 more rows
#> 
#> $valid
#> [1] FALSE
#> 
#> $charts
#> $charts$hepexplorer
#> [1] FALSE
#> 
#> $charts$safetyhistogram
#> [1] FALSE
#> 
#> $charts$safetyoutlierexplorer
#> [1] FALSE
#> 
#> $charts$safetyshiftplot
#> [1] FALSE
#> 
#> $charts$safetyresultsovertime
#> [1] FALSE
#> 
#> $charts$paneledoutlierexplorer
#> [1] FALSE
#> 
#> $charts$safetydeltadelta
#> [1] FALSE
#> 
#> 
#> $status
#> [1] "1 of 33 checks failed."
#> 

# .$valid is now FALSE