Evaluate a data set against a data standard

Determines whether the required data elements in a data standard are found in a given data frame

evaluateStandard(data, standard, includeFields = TRUE, domain = "labs")

Arguments

data
data.frame to evaluate

standard
standard to evaluate

includeFields
should field level data be evaluated?

domain
data domain. "labs" only for now.

Value

a list describing to what degree the data set matches the data standard. The "match" property describes compliance with the standard as "full", "partial" or "none". The "checks" property is a list of the data elements expected for the standard and whether they are "valid" in the given data set. "total_checks", "valid_checks" and "invalid_checks" provide counts of the specified checks. "match_percent" is calculated as valid_checks/total_checks.

Examples

safetyGraphics:::evaluateStandard(data=adlbc, standard="adam") # Match is TRUE

#> $standard
#> [1] "adam"
#> 
#> $checks
#> Source: local data frame [12 x 4]
#> Groups: <by row>
#> 
#> # A tibble: 12 x 4
#>    text_key            standard_val                     type   valid
#>    <chr>               <chr>                            <chr>  <lgl>
#>  1 id_col              USUBJID                          column TRUE 
#>  2 value_col           AVAL                             column TRUE 
#>  3 measure_col         PARAM                            column TRUE 
#>  4 measure_values--ALT Alanine Aminotransferase (U/L)   field  TRUE 
#>  5 measure_values--AST Aspartate Aminotransferase (U/L) field  TRUE 
#>  6 measure_values--TB  Bilirubin (umol/L)               field  TRUE 
#>  7 measure_values--ALP Alkaline Phosphatase (U/L)       field  TRUE 
#>  8 normal_col_low      A1LO                             column TRUE 
#>  9 normal_col_high     A1HI                             column TRUE 
#> 10 studyday_col        ADY                              column TRUE 
#> 11 visit_col           VISIT                            column TRUE 
#> 12 visitn_col          VISITNUM                         column TRUE 
#> 
#> $total_count
#> [1] 12
#> 
#> $valid_count
#> [1] 12
#> 
#> $invalid_count
#> [1] 0
#> 
#> $match_percent
#> [1] 1
#> 
#> $match
#> [1] "full"
#> 

safetyGraphics:::evaluateStandard(data=adlbc, standard="sdtm") # Match is FALSE

#> $standard
#> [1] "sdtm"
#> 
#> $checks
#> Source: local data frame [12 x 4]
#> Groups: <by row>
#> 
#> # A tibble: 12 x 4
#>    text_key            standard_val                      type   valid
#>    <chr>               <chr>                             <chr>  <lgl>
#>  1 id_col              USUBJID                           column TRUE 
#>  2 value_col           STRESN                            column FALSE
#>  3 measure_col         TEST                              column FALSE
#>  4 measure_values--ALT Aminotransferase, alanine (ALT)   field  FALSE
#>  5 measure_values--AST Aminotransferase, aspartate (AST) field  FALSE
#>  6 measure_values--TB  Total Bilirubin                   field  FALSE
#>  7 measure_values--ALP Alkaline phosphatase (ALP)        field  FALSE
#>  8 normal_col_low      STNRLO                            column FALSE
#>  9 normal_col_high     STNRHI                            column FALSE
#> 10 studyday_col        DY                                column FALSE
#> 11 visit_col           VISIT                             column TRUE 
#> 12 visitn_col          VISITNUM                          column TRUE 
#> 
#> $total_count
#> [1] 12
#> 
#> $valid_count
#> [1] 3
#> 
#> $invalid_count
#> [1] 9
#> 
#> $match_percent
#> [1] 0.25
#> 
#> $match
#> [1] "partial"
#>